Fix #8: Improve invite filter regex
- The invite filter regex will now match on multiple "/" characters (if any) on messages

This was a known-ish bug, but was considered low priority or risk because Discord wouldn't display a usable invite. It was considered to be in the same vein as "join the server at discord (dot) gg / diytech" -- something that could be very hard to filter for. Either way, this is a simple enough patch that I overlooked, so... fix.